+++ react-components/src/components/ProductFinder/Interaction.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import Services from '@src/Services'
+import { analytics } from '@src/Helpers'
 import { capitalize } from '@src/Helpers'
 
 function RadioButtons(props) {
@@ -61,7 +62,7 @@ RadioButtons.propTypes = {
 
 
 export default function Interaction(props) {
-  const { txId, attribute, isItemChoice, processResponse } = props
+  const { txId, attribute, isItemChoice, processResponse, isFirstQuestion, searchQuery } = props
 
   const [value, setValue] = useState()
 
@@ -69,6 +70,14 @@ export default function Interaction(props) {
     if (isItemChoice) {
       processResponse(Services.lookupProduct({ q: value.name }))
     } else {
+      if (isFirstQuestion) {
+        analytics({
+          event: 'addProductPageview',
+          virtualPageUrl: '/add-product-modal/tell-us-more',
+          virtualPageTitle: 'Add Product Modal - Tell Us More',
+          productKeyword: capitalize(searchQuery)
+        })
+      }
       processResponse(
         Services.lookupProductRefine({
           txId,